+++ src/Jazor.Admin/AdminShellContracts.cs
@@ -1,5 +1,3 @@
-using Microsoft.AspNetCore.Components;
-
 namespace Jazor.Admin;
 
 [ECMAScript]